[main] Fixed inconsistent handling of constant false/constant true in simplify.py. The new code is untested, but should at least make sense.
Also set ALLOW_CONFLICTING_EFFECTS to False again by default, although
it is activated in Promela and Rovers (which is, of course, a HACK).
While setting it to True seems to work everywhere, some silly stuff
cannot be filtered out in that setting because the instantiator is not
smart enough to recognize mutexes between different SAS+ variables. For
example, with ALLOW_CONFLICTING_EFFECTS=True, it will not notice that
on(a,a) is unreachable in blocksworld/4-0.